tests: transfer tests, refactor - wip
