+++ __mocks__/mockTransferTxReceipts.ts
@@ -1,0 +1,91 @@
+const txReceiptL2 = {
+  value: "0x0000000000000000000000000000000000000000000000000000000000000000",
+  fee: "0x000000000000000000000000000000000000000000000000000000000000000a",
+  transactionType:
+    "0x0000000000000000000000000000000000000000000000000000000000000001",
+  tokenType:
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+  historicRootBlockNumberL2: [
+    "0x000000000000000000000000000000000000000000000000000000000000000c",
+    "0x000000000000000000000000000000000000000000000000000000000000000c",
+  ],
+  historicRootBlockNumberL2Fee: [
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+  ],
+  tokenId: "0x2c76ba40cadcb31819952e94c3d9b7da7fbf8494d8c9d85cfdb162013a5d9195",
+  ercAddress:
+    "0x2e70b6775388fa491e7ae711d327db161b644cdebbb3712853f04e4598d6eba2",
+  recipientAddress:
+    "0xa221891adbe80b3e9966da5c42a8934d94098296f23a05e6dc56b0f60160368b",
+  commitments: [
+    "0x1b5598e47bed4aaadab046836b02a29571d78ef9aad18c85870ec7eec80881a5",
+    "0x00f507063a014a363fb0561193e29392262bacabf099692f38b02fd75133b3dd",
+  ],
+  nullifiers: [
+    "0x29a6d4386fc0b294d823fe4e86e7c7d55f21f37e3928a07aa5a82fff3daa2bdc",
+    "0x25f1842d14316351e303b0496da5797db1f53b034a5778b5a8f9aecbe21a1bbb",
+  ],
+  commitmentFee: [
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+  ],
+  nullifiersFee: [
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+  ],
+  compressedSecrets: [
+    "0x009e448c6db53891ab4db5437b33706b7075cb57dac18e4455796bd763cd265c",
+    "0x162e36f88cf1a18eef667972420897dc26cc74b813db77e7997824d28f15e1ae",
+  ],
+  proof: [
+    "0x1c0b8ead41d5b5e1de227489083a37c0b525c139e661c994dbab92c52e98b30e",
+    "0x2feaa0a6f9b62c565cce09d9085bba0ea76889e1079404e10bcb319ff9b45e01",
+    "0x06ab238a0505eed6e42782919c5edc6fc0526609c802c701c8dd891f99d5f7e1",
+    "0x06c92cab9b51f9e6309de863ed0a2d2506c9a3fbd2693dd0f90b55d233fb17ca",
+    "0x07921dfbf640e3ecd119509709b01bf0a2880d40c4b24e1e499e9bda7f049956",
+    "0x125b1ab13cda30ba6ab3f7a984c684fa8b479041d64830b66533edb8db6373c2",
+    "0x2253120be094d87234d0260d6c32c7952d1ec26c334e86f271e0a6aff7e182b9",
+    "0x2c21996c2a36e7dd2948cd2bf7756ca23c9cd40b880251544fec91d5e1838ebf",
+  ],
+  transactionHash:
+    "0xe761da077b6b5be0e2fc0a3ada285c20e2bf114108907cc62fe8e5fd0c47dfed",
+};
+
+const txReceipt = {
+  blockHash:
+    "0xa60cf941bbb9fea47a0f881dfa54d050ca97b1d04b9623c665c6490e3c6ae165",
+  blockNumber: 7446761,
+  contractAddress: "",
+  cumulativeGasUsed: 47289,
+  effectiveGasPrice: 20000000000,
+  from: "0x02f979a781260955ee760e92e893938ad1ab8a5e",
+  gasUsed: 47289,
+  logs: [
+    {
+      address: "0x2eEeeC446F92c868220CAB1A2ff3cd77937e6560",
+      topics: [
+        "0xd9364d1faedd45a064f9090dd61ade3de8d1c1fd83caaa8ebdc4b9808f4eb989",
+      ],
+      data: "0x",
+      blockNumber: 7446761,
+      transactionHash:
+        "0xbd5d7972be4a35a15700535a5c8f9c453c0ba1140276a87d8315ecd80bfb4921",
+      transactionIndex: 0,
+      blockHash:
+        "0xa60cf941bbb9fea47a0f881dfa54d050ca97b1d04b9623c665c6490e3c6ae165",
+      logIndex: 0,
+      removed: false,
+      id: "log_546fd432",
+    },
+  ],
+  logsBloom:
+    "0x00000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000040000000000000200000000000000020",
+  status: true,
+  to: "0x2eeeec446f92c868220cab1a2ff3cd77937e6560",
+  transactionHash:
+    "0xbd5d7972be4a35a15700535a5c8f9c453c0ba1140276a87d8315ecd80bfb4921",
+  transactionIndex: 0,
+  type: "0x0",
+};
+
+export const transferReceipts = { txReceipt, txReceiptL2 };
