--- __tests__/unit/transactions/transfer.test.ts
@@ -1,12 +1,0 @@
-describe("Transfer transaction", () => {
-  test.todo("Should fail if client throws a Nightfall error");
-  test.todo(
-    "Should throw an error if an exception is caught when submitting on-chain tx",
-  );
-  test.todo(
-    "Should return an instance of <OnChainTransactionReceipts> when submitting on-chain tx",
-  );
-  test.todo(
-    "Should return an instance of <OffChainTransactionReceipt> when sending off-chain tx",
-  );
-});
